Fixed bug where product prices was synced as int instead of float

+++ src/models/Product.php
@@ -88,7 +88,7 @@ class Product extends Model
 
 	public function setSalesPrice($value)
 	{
-		$this->salesPrice = (int) $value;
+		$this->salesPrice = (float) $value;
 		return $this;
 	}
